fs.exists替换为fs.existsSync

+++ coms/build/create.js
@@ -54,34 +54,33 @@ function from(srcname = 'kugou', destname = appname) {
 }
 module.exports = from;
 function copy(path1, path2) {
-    fs.exists(path2, function (exists) {
-        if (exists) console.error(`目标路径已存在：${path2}`);
-        else fs.exists(path.dirname(path2), function (exists) {
-            if (!exists) console.error("目标路径无法创建！");
-            else fs.exists(path1, function (exists) {
-                if (!exists) console.error(`源路径不存在：${path1}`);
-                else fs.stat(path1, function (error, stats) {
-                    if (error) console.error(`无法读取源路径：${path1}`);
-                    else if (stats.isDirectory()) {
-                        fs.readdir(path1, function (error, names) {
-                            if (error) console.error(`读取目录失败：${path1}`)
-                            else fs.mkdir(path2, function (error) {
-                                if (error) console.error(`创建目录失败：${path2}`);
-                                else names.forEach(function (name) {
-                                    copy(path.join(path1, name), path.join(path2, name));
-                                })
-                            });
-                        });
-                    } else {
-                        fs.readFile(path1, function (error, data) {
-                            if (error) console.error(`读取文件失败：${path1}`);
-                            else fs.writeFile(path2, data, function (error) {
-                                if (error) console.error(`写入文件失败：${path2}`);
-                            });
+    if (fs.existsSync(path2)) {
+        console.error(`目标路径已存在：${path2}`);
+    } else if (!fs.existsSync(path.dirname(path2))) {
+        console.error("目标路径无法创建！");
+    } else if (!fs.existsSync(path1)) {
+        console.error(`源路径不存在：${path1}`);
+    } else {
+        fs.stat(path1, function (error, stats) {
+            if (error) console.error(`无法读取源路径：${path1}`);
+            else if (stats.isDirectory()) {
+                fs.readdir(path1, function (error, names) {
+                    if (error) console.error(`读取目录失败：${path1}`)
+                    else fs.mkdir(path2, function (error) {
+                        if (error) console.error(`创建目录失败：${path2}`);
+                        else names.forEach(function (name) {
+                            copy(path.join(path1, name), path.join(path2, name));
                         })
-                    }
+                    });
                 });
-            });
+            } else {
+                fs.readFile(path1, function (error, data) {
+                    if (error) console.error(`读取文件失败：${path1}`);
+                    else fs.writeFile(path2, data, function (error) {
+                        if (error) console.error(`写入文件失败：${path2}`);
+                    });
+                })
+            }
         });
-    });
+    }
 }
